docs: Se cambia logica de configuracion desde main a config
